+++ src/main/java/ti4/buttons/ButtonListener.java
@@ -164,86 +164,6 @@ public class ButtonListener extends ListenerAdapter {
         } else if (buttonID.startsWith("jmfA_") || buttonID.startsWith("jmfN_")) {
             game.initializeMiltySettings().parseButtonInput(event);
 
-
-
-
-
-
-
-
-  
-
-
-
-
-        } else if (buttonID.startsWith("unstableStep3_")) {
-            ButtonHelperActionCards.resolveUnstableStep3(player, game, event, buttonID);
-        } else if (buttonID.startsWith("spaceUnits_")) {
-            ButtonHelperModifyUnits.spaceLandedUnits(buttonID, event, game, player);
-        } else if (buttonID.startsWith("resetSpend_")) {
-            UnfiledButtonHandlers.resetSpend_(event, player, buttonID, game);
-        } else if (buttonID.startsWith("reinforcements_cc_placement_")) {
-            UnfiledButtonHandlers.reinforcementsCCPlacement(event, game, player, buttonID);
-        } else if (buttonID.startsWith("placeHolderOfConInSystem_")) {
-            UnfiledButtonHandlers.placeHolderOfConInSystem(event, game, player, buttonID);
-        } else if (buttonID.startsWith("greyfire_")) {
-            ButtonHelperFactionSpecific.resolveGreyfire(player, game, buttonID, event);
-        } else if (buttonID.startsWith("concludeMove_")) {
-            ButtonHelperTacticalAction.finishMovingForTacticalAction(player, game, event, buttonID);
-        } else if (buttonID.startsWith("transactWith_") || buttonID.startsWith("resetOffer_")) {
-            UnfiledButtonHandlers.transactWith(event, player, buttonID, game);
-        } else if (buttonID.startsWith("rejectOffer_")) {
-            TransactionHelper.rejectOffer(event, player, buttonID, game);
-        } else if (buttonID.startsWith("rescindOffer_")) {
-            TransactionHelper.rescindOffer(event, player, buttonID, game);
-        } else if (buttonID.startsWith("acceptOffer_")) {
-            TransactionHelper.acceptOffer(event, game, player, buttonID);
-        } else if (buttonID.startsWith("play_after_")) {
-            UnfiledButtonHandlers.play_after(event, game, player, buttonID);
-        } else if (buttonID.startsWith("ultimateUndo_")) {
-            UnfiledButtonHandlers.ultimateUndo_(event, game, player, buttonID);
-        } else if (buttonID.startsWith("addIonStorm_")) {
-            ButtonHelper.addIonStorm(game, buttonID, event, player);
-        } else if (buttonID.startsWith("flipIonStorm_")) {
-            ButtonHelper.flipIonStorm(game, buttonID, event);
-        } else if (buttonID.startsWith("terraformPlanet_")) {
-            ButtonHelperFactionSpecific.terraformPlanet(player, buttonID, event, game);
-        } else if (buttonID.startsWith("automatonsPlanet_")) {// "bentorPNPlanet_"
-            ButtonHelperFactionSpecific.automatonsPlanet(buttonID, event, game);
-        } else if (buttonID.startsWith("bentorPNPlanet_")) {// "bentorPNPlanet_"
-            ButtonHelperFactionSpecific.bentorPNPlanet(buttonID, event, game);
-        } else if (buttonID.startsWith("gledgeBasePlanet_")) {
-            ButtonHelperFactionSpecific.gledgeBasePlanet(buttonID, event, game);
-        } else if (buttonID.startsWith("veldyrAttach_")) {
-            ButtonHelperFactionSpecific.resolveBranchOffice(buttonID, event, game, player);
-        } else if (buttonID.startsWith("nanoforgePlanet_")) {
-            UnfiledButtonHandlers.nanoforgePlanet(event, buttonID, game);
-        } else if (buttonID.startsWith("resolvePNPlay_")) {
-            UnfiledButtonHandlers.resolvePNPlay(event, player, buttonID, game);
-        } else if (buttonID.startsWith("send_")) {
-            UnfiledButtonHandlers.send(event, player, buttonID, game);
-        } else if (buttonID.startsWith("replacePDSWithFS_")) {
-            ButtonHelperFactionSpecific.replacePDSWithFS(buttonID, event, game, player);
-        } else if (buttonID.startsWith("putSleeperOnPlanet_")) {
-            ButtonHelperAbilities.putSleeperOn(buttonID, event, game, player);
-        } else if (buttonID.startsWith("frankenDraftAction;")) {
-            FrankenDraftHelper.resolveFrankenDraftAction(game, player, event, buttonID);
-        } else if (buttonID.startsWith("presetEdynAgentStep1")) {
-            ButtonHelperAgents.presetEdynAgentStep1(game, player);
-        } else if (buttonID.startsWith("presetEdynAgentStep2_")) {
-            ButtonHelperAgents.presetEdynAgentStep2(game, player, event, buttonID);
-        } else if (buttonID.startsWith("presetEdynAgentStep3_")) {
-            ButtonHelperAgents.presetEdynAgentStep3(game, player, event, buttonID);
-        } else if (buttonID.startsWith("removeSleeperFromPlanet_")) {
-            ButtonHelperAbilities.removeSleeper(buttonID, event, game, player);
-        } else if (buttonID.startsWith("replaceSleeperWith_")) {
-            ButtonHelperAbilities.replaceSleeperWith(buttonID, event, game, player);
-        } else if (buttonID.startsWith("relicSwapStep2")) {
-            ButtonHelperHeroes.resolveRelicSwapStep2(player, game, event, buttonID);
-        } else if (buttonID.startsWith("relicSwapStep1")) {
-            ButtonHelperHeroes.resolveRelicSwapStep1(player, game, event, buttonID);
-        } else if (buttonID.startsWith("retrieveAgenda_")) {
-            UnfiledButtonHandlers.retrieveAgenda(event, player, buttonID, game);
         } else {
             switch (buttonID) { // TODO Convert all switch case to use @ButtonHandler
                 case "resolveSeizeArtifactStep1" -> ButtonHelperActionCards.resolveSeizeArtifactStep1(player, game, event, "no");
