+++ src/main/java/ti4/commands/game/Undo.java
@@ -3,20 +3,28 @@ package ti4.commands.game;
 import java.io.File;
 import java.util.Arrays;
 import java.util.Comparator;
+import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
+import net.dv8tion.jda.api.entities.channel.concrete.ThreadChannel;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+import net.dv8tion.jda.api.interactions.components.ActionRow;
+import net.dv8tion.jda.api.interactions.components.ItemComponent;
+import net.dv8tion.jda.api.interactions.components.buttons.Button;
 import ti4.helpers.ButtonHelper;
 import ti4.helpers.Constants;
 import ti4.helpers.Storage;
+import ti4.listeners.annotations.ButtonHandler;
 import ti4.map.Game;
 import ti4.map.GameManager;
 import ti4.map.GameSaveLoadManager;
+import ti4.map.Player;
 import ti4.message.MessageHelper;
 
 public class Undo extends GameSubcommandData {
@@ -101,6 +109,99 @@ public class Undo extends GameSubcommandData {
         GameManager.getInstance().deleteGame(game.getName());
         GameManager.getInstance().addGame(gameToRestore);
         GameSaveLoadManager.undo(gameToRestore, event);
+    }
+
+    @ButtonHandler("ultimateUndo")
+    public static void ultimateUndo(ButtonInteractionEvent event, Game game, Player player, String buttonID) {
+        if (game.getSavedButtons().size() > 0 && !game.getPhaseOfGame().contains("status")) {
+            String buttonString = game.getSavedButtons().get(0);
+            if (game.getPlayerFromColorOrFaction(buttonString.split(";")[0]) != null) {
+                boolean showGame = false;
+                for (String buttonString2 : game.getSavedButtons()) {
+                    if (buttonString2.contains("Show Game")) {
+                        showGame = true;
+                        break;
+                    }
+                }
+                if (player != game.getPlayerFromColorOrFaction(buttonString.split(";")[0])
+                    && !showGame) {
+                    MessageHelper.sendMessageToChannel(event.getChannel(),
+                        "You were not the player who pressed the latest button. Use /game undo if you truly want to undo "
+                            + game.getLatestCommand());
+                    return;
+                }
+            }
+        }
+
+        GameSaveLoadManager.undo(game, event);
+        if ("action".equalsIgnoreCase(game.getPhaseOfGame())
+            || "agendaVoting".equalsIgnoreCase(game.getPhaseOfGame())) {
+            if (!event.getMessage().getContentRaw().contains(player.getFinsFactionCheckerPrefix())) {
+                boolean dontDelete = false;
+                for (ActionRow row : event.getMessage().getActionRows()) {
+                    List<ItemComponent> buttonRow = row.getComponents();
+                    for (ItemComponent item : buttonRow) {
+                        if (item instanceof Button butt) {
+                            if (butt.getId().contains("doneLanding")
+                                || butt.getId().contains("concludeMove")) {
+                                dontDelete = true;
+                                break;
+                            }
+                        }
+                    }
+
+                }
+                if (!dontDelete) {
+                    ButtonHelper.deleteMessage(event);
+                }
+            }
+        }
+    }
+
+    @ButtonHandler("ultimateUndo_")
+    public static void ultimateUndo_(ButtonInteractionEvent event, Game game, Player player, String buttonID) {
+        if (!game.getSavedButtons().isEmpty()) {
+            String buttonString = game.getSavedButtons().get(0);
+            String colorOrFaction = buttonString.split(";")[0];
+            Player p = game.getPlayerFromColorOrFaction(colorOrFaction);
+            if (p != null && player != p && !colorOrFaction.equals("null")) {
+                // if the last button was pressed by a non-faction player, allow anyone to undo
+                // it
+                String msg = "You were not the player who pressed the latest button. Use /game undo if you truly want to undo "
+                    + game.getLatestCommand();
+                MessageHelper.sendMessageToChannel(event.getChannel(), msg);
+                return;
+            }
+        }
+        String highestNumBefore = buttonID.split("_")[1];
+        File mapUndoDirectory = Storage.getMapUndoDirectory();
+        if (!mapUndoDirectory.exists()) {
+            return;
+        }
+        String mapName = game.getName();
+        String mapNameForUndoStart = mapName + "_";
+        String[] mapUndoFiles = mapUndoDirectory.list((dir, name) -> name.startsWith(mapNameForUndoStart));
+        if (mapUndoFiles != null && mapUndoFiles.length > 0) {
+            List<Integer> numbers = Arrays.stream(mapUndoFiles)
+                .map(fileName -> fileName.replace(mapNameForUndoStart, ""))
+                .map(fileName -> fileName.replace(Constants.TXT, ""))
+                .map(Integer::parseInt).toList();
+            int maxNumber = numbers.isEmpty() ? 0 : numbers.stream().mapToInt(value -> value).max().orElseThrow(NoSuchElementException::new);
+            if (highestNumBefore.equalsIgnoreCase((maxNumber) + "")) {
+                ButtonHelper.deleteMessage(event);
+            }
+        }
+
+        GameSaveLoadManager.undo(game, event);
+
+        String msg = "You undid something, the details of which can be found in the undo-log thread";
+        List<ThreadChannel> threadChannels = game.getMainGameChannel().getThreadChannels();
+        for (ThreadChannel threadChannel_ : threadChannels) {
+            if (threadChannel_.getName().equals(game.getName() + "-undo-log")) {
+                msg += ": " + threadChannel_.getJumpUrl();
+            }
+        }
+        event.getHook().sendMessage(msg).setEphemeral(true).queue();
     }
 
     public static Map<String, Game> getAllUndoSavedGames(Game game) {
