+++ src/main/java/ti4/helpers/FrankenDraftHelper.java
@@ -19,6 +19,7 @@ import ti4.draft.DraftBag;
 import ti4.draft.DraftItem;
 import ti4.generator.Mapper;
 import ti4.generator.PositionMapper;
+import ti4.listeners.annotations.ButtonHandler;
 import ti4.map.Game;
 import ti4.map.GameSaveLoadManager;
 import ti4.map.Player;
@@ -51,6 +52,7 @@ public class FrankenDraftHelper {
         return buttons;
     }
 
+    @ButtonHandler("frankenDraftAction")
     public static void resolveFrankenDraftAction(Game game, Player player, ButtonInteractionEvent event, String buttonID) {
         String action = buttonID.split(";")[1];
         BagDraft draft = game.getActiveBagDraft();
